Add js functionality for number sections
This commit is where I created 3 wrappers that each contain a
function to toggle my color change functionality. They also contain
a function that cycles through the string or number, flashes however
long the string or number is, and then disappears and has a new section
pop up in its place.

This commit finishes by adding the final event listeners to
colors 2 and 3.