Create the random fortunes and have them appear
I created a placeholder for the fortune in my html. Then I created
a paragraph in js to go in my html placeholder. I appended the
paragraph to the placeholder. After that I created an empty text
node to hold the random fortune. Lastly, I appended the random fortune
to the empty text node.